+++ src/gameObject/PlasmaProjectile.java
@@ -2,7 +2,7 @@ package gameObject;
 
 public class PlasmaProjectile extends Projectile{
 	
-	PlasmaProjectile(int x, int y, Enemy target){
+	PlasmaProjectile(float x, float y, Enemy target){
 		super(x, y, target);
 	}
