+++ src/gameObject/Projectile.java
@@ -3,7 +3,7 @@ package gameObject;
 public abstract class Projectile extends GameObject{
 	Enemy target;
 	
-	Projectile(int x, int y, Enemy target){
+	Projectile(float x, float y, Enemy target){
 		super(x, y);
 		this.target = target;
 	}
